# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ATARK API",
    "version": "1.0.0",
    "description": "API lab3."
  },
  "paths": {
...
    "/api/admin/refresh": {
      "get": {
        "summary": "Token update",
        "tags": ["Admin"],

        "parameters": [
          {
            "name": "refreshToken",
            "in": "body",
            "required": true,
            "schema": { "type": "string" },
            "description": "User email"
          }
        ],
        "responses": {
          "200": {
            "description": "Token updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "refreshToken": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized access" }
        }
      }
    },
...
    "/api/registration": {
      "post": {
        "summary": "User registration",
        "description": "Registration of a new user with email and password validation.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 32,
                    "description": "User password"
                  },
                  "role": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Пользователь успешно зарегистрирован."
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Неверный email или пароль."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "User Login",
        "description": "Login using email and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email"
                  },
                  "password": {
                    "type": "string",
                    "description": "User password"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The user has successfully logged in."
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "access-token-example"
                    },
                    "refreshToken": {
                      "type": "string",
                      "example": "refresh-token-example"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неверный логин или пароль",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Неверный email или пароль."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/activate/{link}": {
      "get": {
        "summary": "Account activation",
        "parameters": [
          {
            "name": "link",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Activation link"
          }
        ],
        "responses": {
          "200": { "description": "Account activated" },
          "400": { "description": "Activation error" }
        }
      }
    }
  },
...
}